+++ eras/conway/test-suite/test/Tests.hs
@@ -7,6 +7,7 @@ import Cardano.Ledger.Conway (Conway)
 import Data.Proxy (Proxy (..))
 import qualified Test.Cardano.Ledger.Babbage.TxInfo as Babbage (txInfoTests)
 import qualified Test.Cardano.Ledger.Conway.Serialisation.CDDL as CDDL
+import qualified Test.Cardano.Ledger.Conway.Serialisation.Failure as Failure
 import qualified Test.Cardano.Ledger.Conway.Serialisation.Roundtrip as Roundtrip
 import qualified Test.Cardano.Ledger.Conway.TxInfo as Conway (txInfoTests)
 import Test.Tasty (TestTree, defaultMain, testGroup)
@@ -22,4 +23,5 @@ defaultTests =
     , CDDL.tests 5
     , Babbage.txInfoTests (Proxy @Conway)
     , Conway.txInfoTests (Proxy @Conway)
+    , Failure.tests
     ]